// types/typo.ts
export type TTextColors =
  | 'text-black'
  | 'text-white'
  | 'text-gray'
  | 'text-lightGray'
  | 'text-orange'
  | 'text-orange1'
  | 'text-yellow'
  | 'text-violet'
  | 'text-pink'
  | 'text-blue'
  | 'text-lightMain'
  | 'text-cancelled'
  | 'text-gray100'
  | 'text-gray200'
  | 'text-gray600'
  | 'text-menuBlack'

export interface ILocation {
  lat?: number | string
  lng?: number | string
}